+++ app/api/counties/__init__.py
@@ -1,0 +1,3 @@
+from .router import router
+
+router = router
