+++ app/api/counties/dependencies.py
@@ -1,0 +1,1 @@
+from app.api.dependencies import get_db
