+++ app/api/counties/router.py
@@ -1,0 +1,12 @@
+from fastapi import APIRouter, Depends
+from sqlalchemy.orm import Session
+
+from .dependencies import get_db
+from .service import get_county_ids
+
+router = APIRouter()
+
+
+@router.get("/ids")
+def read_county_ids(database: Session = Depends(get_db)):
+    return get_county_ids(database)
