+++ app/api/router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from . import user, stores, auth, users, items, districts
+from . import user, stores, auth, users, items, districts, counties
 
 
 api_router = APIRouter()
@@ -10,4 +10,5 @@ api_router.include_router(users.router, prefix="/users")
 api_router.include_router(stores.router, prefix="/stores")
 api_router.include_router(auth.router, prefix="/auth")
 api_router.include_router(items.router, prefix="/items")
+api_router.include_router(counties.router, prefix="/counties")
 api_router.include_router(districts.router, prefix="/districts")
