+++ app/api/stores/models.py
@@ -4,8 +4,9 @@ from sqlalchemy import CheckConstraint, Column, ForeignKey, Integer, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
+from app.api.counties.models import County
+from app.api.districts.models import District
 from app.api.items.models import Item
-from app.api.districts.models import County, District
 from app.database import Base
 from app.utils.validator import cellphone_number_regex, email_regex, telephone_number_regex
